src/main/java/io/lenur/tree/Tree.java
package io.lenur.tree;

public class Tree {
    public boolean isMirror(Node node1, Node node2) {
        if (node1 == null && node2 == null) {
            return true;
        }

        if (node1 == null || node2 == null) {
            return false;
        }

        return node1.val == node2.val
                && isMirror(node1.left, node2.right)
                && isMirror(node1.right, node2.left);
    }

    public boolean isEqual(Node node1, Node node2) {
        if (node1 == null && node2 == null) {
            return true;
        }

        if (node1 == null || node2 == null) {
            return false;
        }

        return node1.val == node2.val
                && isEqual(node1.left, node2.left)
                && isEqual(node1.right, node2.right);
    }
}
